requirements-deploy.txt
# ViFake Analytics — Lightweight deploy requirements
# Dùng file này cho Render / Railway / Koyeb (free tier, RAM <= 512MB)
# Rule-based NLP fallback hoạt động không cần torch/transformers
# XGBoost fusion load từ file .joblib nếu có, fallback sang rule-based nếu không

# === Web Framework ===
fastapi==0.104.0
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# === HTTP ===
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.0
aiofiles>=23.0

# === Authentication & Security ===
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# === Monitoring ===
structlog==23.2.0

# === Utilities ===
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# === Lightweight ML (no torch) ===
numpy>=1.24.0
pandas>=2.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
Pillow>=10.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
